Update old files.
Updating all the old files with author names, and renaming files to match my setup so the next few commits are streamlined. Does not include the "intro" folder because that will be reserved for the next commit. Also I removed "effects" from the weapons since that's gonna be in a seperate XML.