Intro Folder
Merged title_song to titlescreen and credited both author's. Following Clownacy's discovery of the intro state handler when booting up the game, these hacks are numbered 0-4 to match the state # as defined in "0_exe_properties". Some stuff moved around so it appears in the correct XML. I didn't delete anything you just have to find it in another file.