Bitmap Folder
These will be referenced in some other XML's in the future commits. Basically the hackinator doesn't display .bmp files but they still show up in the file_pane, making it super super ugly if you put the bmp with the corresponding xml in the same folder. Working with this limitation, whenever a hack references a bitmap I’m putting it in this folder so that users aren’t clicking on files in other folders and nothing shows up, making it seem broken. It’s hard to explain.