Misc Patches Folder
Oof, this is a big one. The "obsolete" folder is basically XML's that exist but wouldn't recommend because they are kinda absurd or are being replaced with bigger better versions in the near-future, or otherwise are meant to somewhat break the game in interesting ways. Up to you on what you want to include in these folders, I'm just including everything that exists. I.e. someone asked me to undo the multisave so I wrote a xml, etc.